test fixed, files made
