+++ main.go
@@ -1,26 +1,11 @@
 package main
 
 import (
-	"bytes"
 	"database/sql"
-	"encoding/json"
-	"errors"
-	"fmt"
-	"io"
-	"log"
-	"net/http"
-	"os"
-	"path/filepath"
-	"sort"
-	"strconv"
-	"strings"
-	"time"
 
 	_ "github.com/mattn/go-sqlite3"
 	_ "modernc.org/sqlite"
 )
-
-var DBFile = "./scheduler.db"
 
 type Task struct {
 	ID      string `json:"id,omitempty"`
@@ -80,188 +65,7 @@ func isLeapYear(year int) bool {
 	return false
 }
 
-func NextDate(now time.Time, date string, repeat string) (string, error) {
-
-	if repeat == "" {
-		err := errors.New("invalid repeat rule") //дефотлтная ошибка для вывода при некорректном вводе
-		return "", err                           //пустой repeat
-	}
-
-	parsedDate, err := time.Parse("20060102", date)
-	if err != nil {
-		return "", err //неверный формат даты
-	}
-
-	commands := strings.Split(repeat, " ") //вычленяем команды для повторов
-	ruleType := commands[0]                //берем тип команды
-
-	newDate := parsedDate
-	dateFound := false
-	switch ruleType {
-
-	case "d":
-		if len(commands) < 2 {
-			return "", err
-		}
-		repeatPeriod, err := strconv.Atoi(commands[1]) //находим дни для переноса задачи
-
-		if err != nil {
-			return "", err
-		}
-		if repeatPeriod > 400 {
-			err := errors.New("invalid repeat rule") //дефотлтная ошибка для вывода при некорректном вводе
-			return "", err                           //если агрумент для >400
-		}
-		for dateFound == false {
-			if (newDate.After(now) && newDate.After(parsedDate)) || now.Format("20060102") == newDate.Format("20060102") { //если ближайшая дата перевалила за текущий день
-				dateFound = true //ну вот тут вопрос как правильнее, так как при return мы все равно выходим из функции, стоит ли поднимать флаг и стоит ли делать break
-				return newDate.Format("20060102"), nil
-			}
-			newDate = newDate.AddDate(0, 0, repeatPeriod)
-		}
-		return "", err
-
-	case "y":
-		for dateFound == false {
-			if newDate.After(now) && newDate.After(parsedDate) { //если ближайшая дата перевалила за текущий день
-				dateFound = true
-
-				return newDate.Format("20060102"), nil
-			}
-			newDate = newDate.AddDate(1, 0, 0) //находим ближайшую новую дату
-		}
-		return "", err
-
-	case "w":
-
-		if len(commands) < 2 {
-			err := errors.New("invalid repeat rule") //дефолтная ошибка для вывода при некорректном вводе
-			return "", err
-		}
-
-		daysToRepeatStr := commands[1]                           //находим агрумент для повтора при команде повтора на неделю
-		daysToRepeatArray := strings.Split(daysToRepeatStr, ",") //записываем в массив номера дней недели
-		var daysToRepeatInts []int                               //записываем в массив номера дней недели
-
-		for _, day := range daysToRepeatArray {
-
-			dayInt, err := strconv.Atoi(day) //проверяем что день недели корректный
-			if err != nil {
-				return "", err
-			}
-			if dayInt > 7 { // все еще проверяем
-				err := errors.New("invalid repeat rule") //дефолтная ошибка для вывода при некорректном вводе
-				return "", err
-			}
-			daysToRepeatInts = append(daysToRepeatInts, dayInt)
-		}
-
-		sort.Ints(daysToRepeatInts)
-		weekAdd := 0
-		for dateFound == false {
-			sunday := time.Date(newDate.Year(), time.Month(newDate.Month()), newDate.Day()-int(newDate.Weekday())+(weekAdd*7), 0, 0, 0, 0, time.UTC)
-			for _, day := range daysToRepeatInts {
-
-				// Вычисляем конкретный день недели
-
-				newDate := sunday.AddDate(0, 0, day)
-
-				if newDate.After(now) && newDate.After(parsedDate) { //если ближайшая дата перевалила за текущий день
-					dateFound = true
-					return newDate.Format("20060102"), nil
-				}
-			}
-			weekAdd += 1
-		}
-		return "", err
-
-	case "m":
-		daysToRepeatStr := commands[1]                           //получаем аргументы для повторяемых дней
-		daysToRepeatArray := strings.Split(daysToRepeatStr, ",") //переносим дни в массив
-		var daysToRepeatInts []int
-		for _, day := range daysToRepeatArray {
-			dayInt, err := strconv.Atoi(day)
-			if err != nil {
-				return "", err
-			}
-			if dayInt == 0 || dayInt > 31 || dayInt < -2 { // все еще проверяем
-				err := errors.New("invalid repeat rule")
-				return "", err
-			}
-			daysToRepeatInts = append(daysToRepeatInts, dayInt)
-		}
-
-		everyMonthSlice := []int{1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12}
-
-		monthSlice := make([]int, 1, 1)
-
-		if len(commands) > 2 { //если есть аргумент для месяца
-			monthToRepeatStr := commands[2]                            //получаем агрументы для повторяемых месяцев
-			monthToRepeatArray := strings.Split(monthToRepeatStr, ",") // переносим месяцы в массив
-			for _, month := range monthToRepeatArray {
-				monthInt, err := strconv.Atoi(month) //получаем номер месяца из строки
-				if err != nil {
-					return "", err
-				}
-				if monthInt > 12 || monthInt <= 0 {
-					err := errors.New("invalid repeat rule") //дефолтная ошибка для вывода при некорректном вводе
-					return "", err
-				}
-				monthSlice = append(monthSlice, monthInt)
-			}
-		} else {
-			monthSlice = everyMonthSlice
-		}
-		sort.Ints(monthSlice)
-		sort.Ints(daysToRepeatInts)
-
-		year := newDate.Year()
-		for dateFound == false {
-			for _, month := range monthSlice {
-				for _, day := range daysToRepeatInts {
-					newDate = time.Date(year, time.Month(month), day, 0, 0, 0, 0, time.UTC)
-
-					if day < 0 && day > -3 {
-						newDate = time.Date(year, time.Month(month), day+1, 0, 0, 0, 0, time.UTC)
-					}
-					if day == 0 || day < -2 {
-						err := errors.New("invalid repeat rule") //дефолтная ошибка для вывода при некорректном вводе
-						return "", err
-					}
-
-					if newDate.After(now) && newDate.After(parsedDate) { //если ближайшая дата перевалила за текущий день
-						dateFound = true
-						return newDate.Format("20060102"), err
-					}
-				}
-			}
-			year++
-		}
-
-		return "", err
-
-	default:
-		err := errors.New("invalid repeat rule") //дефотлтная ошибка для вывода при некорректном вводе
-		return "", err
-	}
-}
-
-func nextDateHandler(res http.ResponseWriter, req *http.Request) {
-	now := req.FormValue("now")
-	date := req.FormValue("date")
-	repeat := req.FormValue("repeat")
-
-	res.Header().Set("Content-Type", "application/json; charset=UTF-8")
-
-	nowTime, err := time.Parse("20060102", now)
-	if err != nil {
-		return
-	}
-	response, err := NextDate(nowTime, date, repeat)
-	io.WriteString(res, response)
-}
-
-func CreateDB() {
+func CreateDB(DBFile string) {
 	db, err := sql.Open("sqlite3", DBFile)
 
 	if err != nil {
@@ -287,366 +91,6 @@ func CreateDB() {
 	}
 }
 
-func checkInputJSON(res http.ResponseWriter, task *Task) error {
-	var err error
-	res.Header().Set("Content-Type", "application/json; charset=UTF-8")
-	if task.Title == "" {
-		err := errors.New("Title error")
-		response.Error = err.Error()
-		json.NewEncoder(res).Encode(response)
-		return err
-	}
-
-	date := time.Now()
-
-	if task.Date != "" {
-		date, err = time.Parse("20060102", task.Date)
-		if err != nil {
-			err := errors.New("Date error")
-			response.Error = err.Error()
-			json.NewEncoder(res).Encode(response)
-			return err
-		}
-	} else {
-		task.Date = date.Format("20060102")
-	}
-
-	if date.Before(time.Now()) {
-		if task.Repeat == "" {
-			task.Date = time.Now().Format("20060102")
-		} else {
-			task.Date, err = NextDate(time.Now(), task.Date, task.Repeat)
-			if err != nil {
-				err := errors.New("Date error")
-				response.Error = err.Error()
-				json.NewEncoder(res).Encode(response)
-				return err
-			}
-		}
-	}
-	return nil
-}
-
-func taskHandler(res http.ResponseWriter, req *http.Request) {
-	switch req.Method {
-	case http.MethodPost:
-		res.Header().Set("Content-Type", "application/json; charset=UTF-8")
-		var buf bytes.Buffer
-		task := &Task{}
-		_, err := buf.ReadFrom(req.Body)
-		if err != nil {
-			err := errors.New("JSON read error")
-			response.Error = err.Error()
-			json.NewEncoder(res).Encode(response)
-			return
-		}
-
-		err = json.Unmarshal(buf.Bytes(), &task)
-		if err != nil {
-			err := errors.New("JSON Unmarshalling error")
-			response.Error = err.Error()
-			json.NewEncoder(res).Encode(response)
-			return
-		}
-
-		err = checkInputJSON(res, task)
-		if err != nil {
-			err := errors.New("Title error")
-			response.Error = err.Error()
-			json.NewEncoder(res).Encode(response)
-			return
-		}
-
-		db, err := sql.Open("sqlite3", DBFile)
-		defer db.Close()
-		query := "INSERT INTO scheduler (date, title, comment, repeat) VALUES (?, ?, ?, ?)"
-		result, err := db.Exec(query, task.Date, task.Title, task.Comment, task.Repeat)
-
-		if err != nil {
-			return
-		}
-
-		lastId, err := result.LastInsertId()
-		if err != nil {
-			return
-		}
-		response := struct {
-			ID int64 `json:"id"`
-		}{ID: lastId}
-		json.NewEncoder(res).Encode(response)
-
-	case http.MethodGet:
-		id := req.FormValue("id")
-		if id == "" {
-			err := errors.New("Не укаазан ID задачи")
-			response.Error = err.Error()
-			json.NewEncoder(res).Encode(response)
-			return
-		}
-
-		db, err := sql.Open("sqlite3", DBFile)
-		defer db.Close()
-		if err != nil {
-			return
-		}
-
-		query := "SELECT * FROM scheduler WHERE id=?"
-		row, err := db.Query(query, id)
-
-		if err != nil {
-			response.Error = err.Error()
-			json.NewEncoder(res).Encode(response)
-			return
-
-		}
-		task := &Task{}
-		for row.Next() {
-
-			err := row.Scan(&task.ID, &task.Date, &task.Title, &task.Comment, &task.Repeat)
-			if err != nil {
-				response.Error = err.Error()
-				json.NewEncoder(res).Encode(response)
-				return
-			}
-		}
-
-		data, err := json.Marshal(&task)
-		if err != nil {
-			response.Error = err.Error()
-			json.NewEncoder(res).Encode(response)
-			return
-		}
-		res.WriteHeader(http.StatusOK)
-		res.Write(data)
-
-	case http.MethodPut:
-		res.Header().Set("Content-Type", "application/json; charset=UTF-8")
-		var buf bytes.Buffer
-		task := &Task{}
-		_, err := buf.ReadFrom(req.Body)
-		if err != nil {
-			http.Error(res, err.Error(), http.StatusBadRequest)
-			return
-		}
-
-		err = json.Unmarshal(buf.Bytes(), &task)
-		if err != nil {
-			http.Error(res, err.Error(), http.StatusBadRequest)
-			responseError := struct {
-				Error string `json:"error"`
-			}{Error: err.Error()}
-			json.NewEncoder(res).Encode(responseError)
-			return
-		}
-
-		err = checkInputJSON(res, task)
-		if err != nil {
-			response.Error = err.Error()
-			json.NewEncoder(res).Encode(response)
-			return
-		}
-
-		db, err := sql.Open("sqlite3", DBFile)
-		defer db.Close()
-		query := "UPDATE scheduler SET date = ?, title = ?, comment = ?, repeat = ? WHERE id = ?"
-		_, err = db.Exec(query,
-			task.Date,
-			task.Title,
-			task.Comment,
-			task.Repeat,
-			task.ID)
-
-		if err != nil {
-			response.Error = err.Error()
-			json.NewEncoder(res).Encode(response)
-			return
-		}
-		json.NewEncoder(res).Encode(response)
-		return
-
-	case http.MethodDelete:
-		id := req.FormValue("id")
-		if id == "" {
-			err := errors.New("Не укаазан ID задачи")
-			response.Error = err.Error()
-			json.NewEncoder(res).Encode(response)
-			return
-		}
-
-		db, err := sql.Open("sqlite3", DBFile)
-		defer db.Close()
-		if err != nil {
-			return
-		}
-
-		query := "DELETE FROM scheduler WHERE id=?"
-		_, err = db.Exec(query, id)
-
-		if err != nil {
-			response.Error = err.Error()
-			json.NewEncoder(res).Encode(response)
-			return
-
-		}
-		json.NewEncoder(res).Encode(response)
-	}
-}
-
-func doneTask(res http.ResponseWriter, req *http.Request) {
-	switch req.Method {
-	case http.MethodPost:
-		id := req.FormValue("id")
-		if id == "" {
-			err := errors.New("Не укаазан ID задачи")
-			response.Error = err.Error()
-			json.NewEncoder(res).Encode(response)
-			return
-		}
-
-		db, err := sql.Open("sqlite3", DBFile)
-		defer db.Close()
-		if err != nil {
-			return
-		}
-
-		query := "SELECT * FROM scheduler WHERE id=?"
-		row, err := db.Query(query, id)
-
-		if err != nil {
-			response.Error = err.Error()
-			json.NewEncoder(res).Encode(response)
-			return
-
-		}
-		task := &Task{}
-		for row.Next() {
-
-			err := row.Scan(&task.ID, &task.Date, &task.Title, &task.Comment, &task.Repeat)
-			if err != nil {
-				response.Error = err.Error()
-				json.NewEncoder(res).Encode(response)
-				return
-			}
-		}
-		if task.Repeat == "" {
-			query := "DELETE FROM scheduler WHERE id=?"
-			_, err := db.Exec(query, id)
-
-			if err != nil {
-				response.Error = err.Error()
-				json.NewEncoder(res).Encode(response)
-				return
-			}
-		} else {
-			newDate, err := NextDate(time.Now(), task.Date, task.Repeat)
-			if err != nil {
-				response.Error = err.Error()
-				json.NewEncoder(res).Encode(response)
-				return
-			}
-			query := "UPDATE scheduler SET date=? WHERE id=?"
-			_, err = db.Exec(query, newDate, id)
-
-			if err != nil {
-				response.Error = err.Error()
-				json.NewEncoder(res).Encode(response)
-				return
-			}
-		}
-		_, err = json.Marshal(response)
-		if err != nil {
-			response.Error = err.Error()
-			json.NewEncoder(res).Encode(response)
-			return
-		}
-		json.NewEncoder(res).Encode(response)
-	}
-}
-
-func getTasks(res http.ResponseWriter, req *http.Request) {
-	switch req.Method {
-	case http.MethodGet:
-		db, err := sql.Open("sqlite3", DBFile)
-		defer db.Close()
-
-		if err != nil {
-			http.Error(res, err.Error(), http.StatusBadRequest)
-			response.Error = err.Error()
-			json.NewEncoder(res).Encode(response)
-			return
-		}
-
-		rows, err := db.Query("SELECT * FROM scheduler ORDER BY date LIMIT ?", 10)
-		if err != nil {
-			response.Error = err.Error()
-			json.NewEncoder(res).Encode(response)
-			return
-		}
-		tasks := &Tasks{}
-		for rows.Next() {
-			task := &Task{}
-
-			err := rows.Scan(&task.ID, &task.Date, &task.Title, &task.Comment, &task.Repeat)
-			if err != nil {
-				response.Error = err.Error()
-				json.NewEncoder(res).Encode(response)
-				return
-			}
-			tasks.Tasks = append(tasks.Tasks, *task)
-		}
-		if tasks.Tasks == nil {
-			tasks.Tasks = []Task{}
-		}
-		data, err := json.Marshal(&tasks)
-		if err != nil {
-			response.Error = err.Error()
-			json.NewEncoder(res).Encode(response)
-			return
-		}
-		res.WriteHeader(http.StatusOK)
-		res.Write(data)
-	}
-}
 func main() {
-	var webDir = "./web"
-
-	var port = ":7540"
-	if len(os.Getenv("TODO_PORT")) > 0 {
-		port = os.Getenv("TODO_PORT")
-	}
-
-	appPath, err := os.Executable()
-	if err != nil {
-		log.Fatal(err)
-	}
-
-	if len(os.Getenv("TODO_DBFILE")) > 0 {
-		DBFile = filepath.Join(filepath.Dir(appPath), "TODO_DBFILE")
-	}
-
-	_, err = os.Stat(DBFile)
-
-	var install bool
-	if err != nil {
-		install = true
-	}
-	if install {
-		CreateDB()
-	}
-
-	http.Handle("/", http.FileServer(http.Dir(webDir)))
-
-	http.HandleFunc("/api/task", taskHandler)
-	http.HandleFunc("/api/task/done", doneTask)
-
-	http.HandleFunc("/api/tasks", getTasks)
-
-	http.HandleFunc("/api/nextdate", nextDateHandler)
-
-	err = http.ListenAndServe(port, nil)
-	if err != nil {
-		panic(err)
-	}
-	fmt.Println("Завершаем работу")
-
+	serverStart()
 }
