+++ server.go
@@ -1,0 +1,51 @@
+package main
+
+import (
+	"fmt"
+	"log"
+	"net/http"
+	"os"
+	"path/filepath"
+)
+
+func serverStart() {
+
+	if len(os.Getenv("TODO_PORT")) > 0 {
+		port = os.Getenv("TODO_PORT")
+	}
+
+	appPath, err := os.Executable()
+	if err != nil {
+		log.Fatal(err)
+	}
+
+	if len(os.Getenv("TODO_DBFILE")) > 0 {
+		DBFile = filepath.Join(filepath.Dir(appPath), "TODO_DBFILE")
+	}
+
+	_, err = os.Stat(DBFile)
+
+	var install bool
+	if err != nil {
+		install = true
+	}
+	if install {
+		CreateDB(DBFile)
+	}
+
+	http.Handle("/", http.FileServer(http.Dir(webDir)))
+
+	http.HandleFunc("/api/task", taskHandler)
+	http.HandleFunc("/api/task/done", doneTask)
+
+	http.HandleFunc("/api/tasks", getTasks)
+
+	http.HandleFunc("/api/nextdate", nextDateHandler)
+
+	err = http.ListenAndServe(port, nil)
+	if err != nil {
+		panic(err)
+	}
+	fmt.Println("Завершаем работу")
+	return
+}
