+++ settings.go
@@ -1,0 +1,6 @@
+package main
+
+var DBFile = "./scheduler.db"
+var webDir = "./web"
+
+var port = ":7540"
