+++ tests/addtask_4_test.go
@@ -149,7 +149,7 @@ func TestAddTask(t *testing.T) {
 				continue
 			}
 			if today && task.Date != now.Format(`20060102`) {
-				t.Errorf("Дата должна быть сегодняшняя %v", v) //Чего? В nextdate_3_test.go при правиле d 1 ожидается завтрашняя дата, а тут мы сегодняшнюю ждем
+				t.Errorf("Дата должна быть сегодняшняя %v", v)
 			}
 		}
 	}
@@ -159,7 +159,7 @@ func TestAddTask(t *testing.T) {
 		{"20231220", "Сделать что-нибудь", "Хорошо отдохнуть", ""},
 		{"20240108", "Уроки", "", "d 10"},
 		{"20240102", "Отдых в Сочи", "На лыжах", "y"},
-		//{"today", "Фитнес", "", "d 1"}, //коммент в проверке к этой записи
+		{"today", "Фитнес", "", "d 1"},
 		{"today", "Шмитнес", "", ""},
 	}
 	check()
