+++ tests/nextdate_3_test.go
@@ -35,7 +35,7 @@ func TestNextDate(t *testing.T) {
 		{"20240202", "d 30", `20240303`},
 		{"20240320", "d 401", ""},
 		{"20231225", "d 12", `20240130`},
-		{"20240228", "d 1", "20240229"}, //коммент в проверке к этой записи в addtask_4_test.go
+		{"20240228", "d 1", "20240229"},
 	}
 	check := func() {
 		for _, v := range tbl {
@@ -61,7 +61,7 @@ func TestNextDate(t *testing.T) {
 		{"20240120", "m 40,11,19", ""},
 		{"20240116", "m 16,5", "20240205"},
 		{"20240126", "m 25,26,7", "20240207"},
-		//{"20240409", "m 31", "20240531"},
+		{"20240409", "m 31", "20240531"},
 		{"20240329", "m 10,17 12,8,1", "20240810"},
 		{"20230311", "m 07,19 05,6", "20240507"},
 		{"20230311", "m 1 1,2", "20240201"},
@@ -70,7 +70,6 @@ func TestNextDate(t *testing.T) {
 		{"20240222", "m -2,-3", ""},
 		{"20240326", "m -1,-2", "20240330"},
 		{"20240201", "m -1,18", "20240218"},
-		////////////////////////////////////
 		{"20240125", "w 1,2,3", "20240129"},
 		{"20240126", "w 7", "20240128"},
 		{"20230126", "w 4,5", "20240201"},
